RAC-186 refactor: `SalaryUtil`에서 정산일 계산
