+++ src/main/java/com/postgraduate/domain/salary/application/usecase/SalaryInfoUseCase.java
@@ -15,11 +15,12 @@ import org.springframework.transaction.annotation.Transactional;
 import java.time.LocalDate;
 import java.util.List;
 
+import static com.postgraduate.domain.salary.util.SalaryUtil.getSalaryDate;
+
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class SalaryInfoUseCase {
-    private static final int SALARY_DATE = 10;
     private final SeniorGetService seniorGetService;
     private final SalaryGetService salaryGetService;
 
@@ -29,13 +30,6 @@ public class SalaryInfoUseCase {
         List<Salary> salaries = salaryGetService.bySeniorAndSalaryDate(senior, salaryDate);
         int amount = getAmount(salaries);
         return new SalaryInfoResponse(salaryDate, amount); //TODO 수수료
-    }
-
-    private LocalDate getSalaryDate() {
-        LocalDate now = LocalDate.now();
-        return now.getDayOfMonth() <= SALARY_DATE
-                ? now.withDayOfMonth(SALARY_DATE)
-                : now.plusMonths(1).withDayOfMonth(SALARY_DATE);
     }
 
     private int getAmount(List<Salary> salaries) {
