--- src/main/java/com/postgraduate/domain/salary/util/MonthFormat.java
@@ -1,9 +1,0 @@
-package com.postgraduate.domain.salary.util;
-
-import java.time.format.DateTimeFormatter;
-
-public class MonthFormat {
-    public static DateTimeFormatter getMonthFormat() {
-        return DateTimeFormatter.ofPattern("yyyy-MM");
-    }
-}
